daneel_package/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "daneel"
version = "0.0.1"
authors = [
  { name="Giacomo Menegatti", email="giacomo.menegatti@studenti.unipd.it" },
]
description = "Daneel detection package"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["numpy", "matplotlib", "batman-package"]


classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
